Add support for a slide querystring argument
This commit handles setting an optional `slide_arg` option. If set, the query-string value mapping to this key is retrieved and used to seed the first slide shown.

The change also properly handles setting the slide to a negative number, or a number larger than the number of slides by checking against the start and end option bounds. 

Note that we have to animate the container in because otherwise setting a non-0 slide as the current slide would produce a container that was hidden, even though there are contents.